refactor: cleaner code
